% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaTOpathway.r
\name{NameMappingExact}
\alias{NameMappingExact}
\title{A NameMappingExact Function}
\usage{
NameMappingExact()
}
\description{
This function allows you to exact matching of your metabolites names with data base
consists of more than 20k metabolites. it allows you also to search among metabloites synanoms.
make sure that you have the below two databases for the exact matching:
cmpd_db.rda, and syn_nms_db.rda
This function was modified version of the name.match function in the below link:
https://github.com/cangfengzhe/Metabo/blob/master/MetaboAnalyst/website/name_match.R
}
\examples{
NameMappingExact()
NameMappingExact()


}
\keyword{Exact}
\keyword{Matching}
